Implement overworld expansion

+++ src/fame_checker.c
@@ -288,7 +288,7 @@ static const u8 *const sFameCheckerFlavorTextPointers[] = {
     gFameCheckerFlavorText_Giovanni0, gFameCheckerFlavorText_Giovanni1, gFameCheckerFlavorText_Giovanni2, gFameCheckerFlavorText_Giovanni3, gFameCheckerFlavorText_Giovanni4, gFameCheckerFlavorText_Giovanni5
 };
 
-static const u8 sFameCheckerArrayNpcGraphicsIds[] = {
+static const u16 sFameCheckerArrayNpcGraphicsIds[] = {
     // Pallet
     OBJ_EVENT_GFX_CUT_TREE,
     OBJ_EVENT_GFX_BLUE,
